Do not alter HOME
gofish do not use HOME for anything other than computing the path of the cache directory containing tarballs. So avoid it, so that it does not incur bad side-effects, like `kubectl` can not access `$HOME/.kube/config` as the HOME is somewhere else after running shoal as a go library.